updated float in scoring

+++ FRC-2025/src/main/java/frc/robot/Subsystems/Scoring.java
@@ -26,7 +26,7 @@ public class Scoring extends SubsystemBase {
         return sensor.get();
     }
 
-    public void setRollerSpeed(float speed) {
+    public void setRollerSpeed(double speed) {
         roller.set(speed);
     }
